+++ CHANGES.md
@@ -21,6 +21,8 @@ This is mimiced in Setup ATM, but will be changed in 2.0.
 
 * Removed `Operation::run` as it was a bad decision. Raising an exception on invalid is a very test-specific scenario and shouldn't have been handled in the core doce.
 * Removed `Operation::present`, since you can simply call `Operation::new` (without builders) or `Operation::build_operation` (with builders).
+* Removed `Operation::valid?`. This is in the result object via `result[:valid]`.
+* Removed `Operation#errors`. This is in the result object via `result[:errors]` if the operation was invalid.
 * Removed the private option `:raise_on_invalid`. Use `Contract::Raise` instead, if you need it in tests.
 
 ## #call